encode datastore credentials if provided
